Initial version of OS Performance Logger
A simple program that logs system and process CPU load together with VM
runtime information to a CSV file.

The purpose of this program is to show that the Java runtime has a bug
that on Windows machines result in that CPU load will forever get stuck
once the JVM process has run for 24.8 days (2147483.647 seconds).